# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CPCBusiness Enterprise Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── CLIENTS ─────────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # ─── TASKS ───────────────────────────────────────────────────────────────────
...
  /portfolio:
    get:
      operationId: getPortfolioItems
      tags: [portfolio]
      summary: List portfolio items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: createPortfolioItem
      tags: [portfolio]
      summary: Create portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
  /portfolio/{id}:
    patch:
      operationId: updatePortfolioItem
      tags: [portfolio]
      summary: Update portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
    delete:
      operationId: deletePortfolioItem
      tags: [portfolio]
      summary: Delete portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── MESSAGES ────────────────────────────────────────────────────────────────
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a conversation
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: recipientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ───────────────────────────────────────────────────────
...
    Lead:
      type: object
      required: [id, name, email, stage, source, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        stage:
          type: string
        source:
          type: string
        value:
          type: ["number", "null"]
        assignedTo:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadInput:
      type: object
      required: [name, email, stage, source]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        stage:
          type: string
        source:
          type: string
        value:
          type: number
        assignedTo:
          type: integer
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        stage:
          type: string
        source:
          type: string
        value:
          type: number
        assignedTo:
          type: integer
        notes:
          type: string

    # ─── CLIENT SCHEMAS ─────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, name, status, priority, startDate, endDate, progress, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
        budget:
          type: ["number", "null"]
        spent:
          type: ["number", "null"]
        startDate:
          type: string
        endDate:
          type: string
        teamMembers:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name, status, priority, startDate, endDate]
      properties:
        name:
          type: string
        description:
          type: string
        clientId:
          type: integer
        status:
          type: string
        priority:
          type: string
        budget:
          type: number
        startDate:
          type: string
        endDate:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
        budget:
          type: number
        spent:
          type: number
        endDate:
          type: string
    ProjectStats:
      type: object
      required: [totalTasks, completedTasks, openTickets, teamSize, daysRemaining]
      properties:
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        openTickets:
          type: integer
        teamSize:
          type: integer
        daysRemaining:
          type: integer

    # ─── TASK SCHEMAS ───────────────────────────────────────────────────────
...
    PortfolioItem:
      type: object
      required: [id, title, category, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        client:
          type: ["string", "null"]
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        results:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
    PortfolioItemInput:
      type: object
      required: [title, category, description]
      properties:
        title:
          type: string
        category:
          type: string
        client:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        imageUrl:
          type: string
        technologies:
          type: array
          items:
            type: string
        results:
          type: string
        liveUrl:
          type: string
        isFeatured:
          type: boolean
    PortfolioItemUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        technologies:
          type: array
          items:
            type: string
        results:
          type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean

    # ─── MESSAGE SCHEMAS ────────────────────────────────────────────────────
    Message:
      type: object
      required: [id, senderId, senderName, content, createdAt]
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderAvatarUrl:
          type: ["string", "null"]
        recipientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        recipientId:
          type: integer
        projectId:
          type: integer

    # ─── NOTIFICATION SCHEMAS ───────────────────────────────────────────────
...