# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPCBusiness Enterprise Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  totalTasks: number;
  completedTasks: number;
  openTickets: number;
  teamSize: number;
  daysRemaining: number;
}